requirements.txt
# requirements.txt

streamlit
google-adk
pypdf2
python-dotenv
